Handle animations with duration: Infinity
Animations with no duration will have a duration of Infinity. The
progress of an infinite animation, anim, is (time.current - anim.start) as
opposed to ((time.current - anim.start) / anim.duration). Infinite
animations with array properties will simply use the first element of
the array of that property therefore, a function callback is more
useful.

The progress of an animation is now also passed to the anim.object.draw
method.